requirements.txt
datasets==2.15.0
pyyaml==6.0.1
matplotlib==3.8.0
numpy==1.26.0
python-dotenv==1.0.0
requests==2.31.0